feat(frontend): split Roles into its own tab, separate from Personas (#1146)
Turn limits are a role setting and must never read as a persona one. The Personas
tab hosted both editors, so the per-role max_turns field sat next to persona
fields and could be misread. Extract the roles registry (name, body, and the
max_turns cap) into a dedicated Roles tab/page; Personas now only reads the role
list to offer routing chips and never edits roles or any turn limit. Roles default
to max_turns -1 (infinite). No backend change.
